Tested the update and the delete flow.
Performed a partial design for the search operation. Isolated the
identity operations into the Identity controller and cleaned the
loginController. Modified links on the adminPage to point to the right
page. Modified a constant in the constants class. Added new pages for
update and delete operations. Updated the web.xml to have more url
patterns to point to the IdentityController.